+++ crawler/types/article.rb
@@ -19,16 +19,28 @@ module Zendesk
       @crawler.get Section, @zendesk_obj.section_id
     end
 
-    protected
-
     def ignore? t
-      ap = @zendesk_obj.section.access_policy['access_policy']['viewable_by']
-      super(t) || @zendesk_obj.draft || t.draft || ap != 'everybody'
+      super(t) ||
+        @zendesk_obj.draft ||
+        t.draft ||
+        !section.exists?(t.locale) ||
+        section.ignore?(t)
     end
 
+    protected
+
     def selected
-      keys = %i(promoted position vote_sum comments_disabled label_names)
+      keys = %i(promoted position vote_sum comments_disabled)
       keys.map { |k| [k, @zendesk_obj[k]] }.to_h
+    end
+
+    def tags locale
+      labels = @zendesk_obj.label_names || []
+      res = labels.grep(/^#{locale}:/).map { |l| l.gsub(/^#{locale}:/, '') }
+      return res unless res.empty?
+      res = labels.grep(/^#{locale[0...2]}:/).map { |l| l.gsub(/^#{locale[0...2]}:/, '') }
+      return res unless res.empty?
+      return labels
     end
 
     def translation t
@@ -39,6 +51,7 @@ module Zendesk
         section: simple_section.merge(
           full_path: "#{simple_category[:title]} > #{simple_section[:title]}"
         ),
+        label_names: tags(t.locale),
         created_at_iso: @zendesk_obj.created_at.utc.iso8601,
         updated_at_iso: @zendesk_obj.updated_at.utc.iso8601
       )